requirements.txt
python-telegram-bot==13.15
openai==1.14.3
